fix addTrigger's max algo - cast idx attr to num
The addTrigger function's max finding algorithm was using a string, as that is what $.attr() returns. Using the unary + operator we can coerce it back into an integer. This allows the algorithm to work properly. Otherwise, it'll think "10" comes before "9". Also, let's actually use the idx variable instead of using $.attr() multiple times.